pyproject.toml
[project]
name = "ml-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn==1.6.1",
]
